# docs/openapi/dpp-api-v1.yaml
openapi: 3.1.0
info:
  title: Claros DPP Public and Integration API
  version: "2026-06-27"
  description: >
    Current external API surface for Digital Product Passports. Public reads use
    `/api/public/passports/{dppId}` and return public fields by default. Selected
    restricted fields are returned only when a valid security group API key is
    supplied in `X-API-Key` or `X-Security-Group-Key`. Company automation writes
    use the company-name slug integration namespace and require a Bearer token.
servers:
  - url: https://www.claros-dpp.online
    description: Production
  - url: http://localhost:3001
    description: Local backend
...
paths:
...
  /api/companies/{companySlug}/integrations/v1/passports:
    post:
      tags: [Integration write]
      summary: Create an editable passport draft for a company
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CompanySlugPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PassportMutationRequest'
      responses:
        '201':
          description: Created DPP draft.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Identifier or lifecycle conflict.
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CompanySlugPath:
      name: companySlug
      in: path
      required: true
      schema:
        type: string
      description: Lowercase hyphenated company name, for example `acme-batteries`.
...
  schemas:
...
    PassportMutationRequest:
      type: object
      additionalProperties: true
      properties:
        passportType:
          type: string
        productIdentifier:
          type: string
        granularity:
          type: string
          enum: [model, batch, item]
        fields:
          type: object
          additionalProperties: true
...
  responses:
    BadRequest:
      description: Invalid request.
    Unauthorised:
      description: Missing or invalid Bearer token.
    Forbidden:
      description: Authenticated caller is not allowed to perform this action.
    NotFound:
      description: Resource not found.
...
    InternalError:
      description: Internal server error.